Move error codes into module (internal change only - public API stays the same).
